be/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-camel-case==1.4.2
mpmath==1.3.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
